Homepage and friendProfile change
Finished search function of homepage so user can look for other users. Implemented suggestions and error handling if the user doesn't exist. Made redirection so page is redirected to the searched user. Renamed friendProfile to userProfile because that sounded more fitting. Still need to have the page not be hard-coded.